Fix SSR with Redux Example
