+++ examples/ssr-with-redux/src/app/actionTypes/CommentsActionType.js
