--- examples/ssr-with-redux/src/app/actionTypes/IndexActionTypes.js
@@ -1,5 +1,0 @@
-const IndexActionTypes = {
-  ADD_ITEMS: Symbol("index/addItems")
-};
-
-export default IndexActionTypes;
