+++ examples/ssr-with-redux/src/app/actionTypes/ItemActionType.js
