--- examples/ssr-with-redux/src/app/actionTypes/ItemActionTypes.js
@@ -1,7 +1,0 @@
-const ItemActionTypes = {
-  NEW_ITEM: Symbol("item/newItem"),
-  ADD_COMMENTS: Symbol("item/addComments"),
-  CLEAR_COMMENTS: Symbol("item/clearComments")
-};
-
-export default ItemActionTypes;
