+++ examples/ssr-with-redux/src/app/actionTypes/ItemsActionType.js
