--- examples/ssr-with-redux/src/app/actionTypes/NewsActionTypes.js
@@ -1,5 +1,0 @@
-const NewsActionTypes = {
-  ADD_ITEMS: Symbol("news/addItems")
-};
-
-export default NewsActionTypes;
