+++ examples/ssr-with-redux/src/app/actionTypes/UserActionType.js
