--- examples/ssr-with-redux/src/app/actionTypes/UserActionTypes.js
@@ -1,5 +1,0 @@
-const UserActionTypes = {
-  NEW_USER: Symbol("user/newUser")
-};
-
-export default UserActionTypes;
