+++ examples/ssr-with-redux/src/app/pages/UserPage.js
@@ -2,7 +2,7 @@ import NProgress from "nprogress";
 import React from "react";
 import { createStore } from "redux";
 import { Provider } from "react-redux";
-import userRootReducer from "../rootReducers/user";
+import userRootReducer from "../reducers/user";
 import UserContainer from "../containers/UserContainer";
 import HackerNewsApi from "../api/HackerNewsApi";
 import NotFoundPage from "./NotFoundPage";
