+++ examples/ssr-with-redux/src/app/reducers/news/items.js
@@ -1,10 +1,10 @@
 import Immutable from "immutable";
 import Item from "../../models/Item";
-import NewsActionTypes from "../../actionTypes/NewsActionTypes";
+import { itemsActionType } from "../../actionTypes/ItemsActionType";
 
 export default function items(state = Immutable.List(), action) {
   switch (action.type) {
-    case NewsActionTypes.ADD_ITEMS:
+    case itemsActionType.SYNC:
       const newState = Immutable.List();
       return newState.withMutations((newState) => {
         action.items.map((itemAttributes) =>
