--- examples/ssr-with-redux/src/app/reducers/user.js
@@ -1,11 +1,0 @@
-import User from "../models/User";
-import { userActionType } from "../actionTypes/UserType";
-
-export default function user(state = null, action) {
-  switch (action.type) {
-    case userActionType.NEW_INSTANCE:
-      return new User(action.attributes);
-    default:
-      return state;
-  }
-}
