--- examples/ssr-with-redux/src/app/rootReducers/item/index.js
@@ -1,10 +1,0 @@
-import { combineReducers } from "redux";
-import comments from "../../reducers/comments";
-import item from "../../reducers/item";
-
-const rootReducer = combineReducers({
-  comments,
-  item
-});
-
-export default rootReducer;
