--- examples/ssr-with-redux/src/app/rootReducers/news/index.js
@@ -1,8 +1,0 @@
-import { combineReducers } from "redux";
-import items from "../../reducers/items";
-
-const rootReducer = combineReducers({
-  items
-});
-
-export default rootReducer;
