--- examples/ssr-with-redux/src/app/rootReducers/user/index.js
@@ -1,8 +1,0 @@
-import { combineReducers } from "redux";
-import user from "../../reducers/user";
-
-const rootReducer = combineReducers({
-  user
-});
-
-export default rootReducer;
